+++ scripts/context-menu/gnome/install-gnome-extension.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-set -eu
-
-cargo run -p linkforge-context-menu-gnome -- install "$@"